fix(mac): complete macOS compilation - fix all min/max/MIN/MAX macro calls and pointer type mismatches
- Add #include <algorithm> to 30+ compilation units
- Replace all standalone min/max/MIN/MAX with std::min/std::max
- Fix pointer-to-int truncation with (intptr_t)/(uintptr_t) casts
- Fix SignedByte*/Byte* type mismatches in xfer calls
- Fix std::max type promotion issues (Real/Int compatibility)
- Files fixed: Object.cpp, Locomotor.cpp, SabotageSupplyCenterCrateCollide.cpp,
  SabotageSupplyDropzoneCrateCollide.cpp, ConvertToHijackedVehicleCrateCollide.cpp,
  TerrainLogic.cpp, HighlanderBody.cpp, ImmortalBody.cpp, UndeadBody.cpp,
  DumbProjectileBehavior.cpp, MinefieldBehavior.cpp, PoisonedBehavior.cpp,
  SlowDeathBehavior.cpp, SabotageInternetCenterCrateCollide.cpp, ActiveBody.cpp, and many behavior modules

**Build Status**: 0 compilation errors, 148 warnings (pre-existing offsetof/sprintf/etc)

Refs: Phase 29.x - Platform Abstraction Layer Completion

+++ GeneralsMD/Code/GameEngine/Source/GameClient/GUI/GUICallbacks/Menus/WOLGameSetupMenu.cpp
@@ -533,7 +533,7 @@ static void handleColorSelection(int index)
 	GameWindow *combo = comboBoxColor[index];
 	Int color, selIndex;
 	GadgetComboBoxGetSelectedPos(combo, &selIndex);
-	color = (Int)GadgetComboBoxGetItemData(combo, selIndex);
+	color = (Int)(intptr_t)GadgetComboBoxGetItemData(combo, selIndex);
 
 	GameInfo *myGame = TheGameSpyInfo->getCurrentStagingRoom();
 
@@ -596,7 +596,7 @@ static void handlePlayerTemplateSelection(int index)
 	GameWindow *combo = comboBoxPlayerTemplate[index];
 	Int playerTemplate, selIndex;
 	GadgetComboBoxGetSelectedPos(combo, &selIndex);
-	playerTemplate = (Int)GadgetComboBoxGetItemData(combo, selIndex);
+	playerTemplate = (Int)(intptr_t)GadgetComboBoxGetItemData(combo, selIndex);
 	GameInfo *myGame = TheGameSpyInfo->getCurrentStagingRoom();
 
 	if (myGame)
@@ -721,7 +721,7 @@ static void handleTeamSelection(int index)
 	GameWindow *combo = comboBoxTeam[index];
 	Int team, selIndex;
 	GadgetComboBoxGetSelectedPos(combo, &selIndex);
-	team = (Int)GadgetComboBoxGetItemData(combo, selIndex);
+	team = (Int)(intptr_t)GadgetComboBoxGetItemData(combo, selIndex);
 	GameInfo *myGame = TheGameSpyInfo->getCurrentStagingRoom();
 
 	if (myGame)
@@ -767,7 +767,7 @@ static void handleStartingCashSelection()
     GadgetComboBoxGetSelectedPos(comboBoxStartingCash, &selIndex);
 
     Money startingCash;
-    startingCash.deposit( (UnsignedInt)GadgetComboBoxGetItemData( comboBoxStartingCash, selIndex ), FALSE, FALSE );
+    startingCash.deposit( (UnsignedInt)(uintptr_t)GadgetComboBoxGetItemData( comboBoxStartingCash, selIndex ), FALSE, FALSE );
     myGame->setStartingCash( startingCash );
     myGame->resetAccepted();
 
@@ -1021,7 +1021,7 @@ void WOLDisplayGameOptions( void )
   Int index = 0;
   for ( ; index < itemCount; index++ )
   {
-    Int value  = (Int)GadgetComboBoxGetItemData(comboBoxStartingCash, index);
+    Int value  = (Int)(intptr_t)GadgetComboBoxGetItemData(comboBoxStartingCash, index);
     if ( value == theGame->getStartingCash().countMoney() )
     {
       // Note: must check if combobox is already correct to avoid infinite recursion
